Enhancement：Refine code.
